requirements.txt
streamlit
newsapi-python
python-dotenv
google-generativeai